+++ Editor/Scripts/NodeEditor.cs
@@ -117,7 +117,7 @@ namespace LunraGames.NoiseMaker
 								{
 									for (var y = 0; y < height; y++)
 									{
-										var value = (float)module.GetValue((double)x, (double)y, 0.0);
+										var value = module.GetValue(x, y, 0f);
 										pixels[SphereUtils.PixelCoordinateToIndex(x, y, width, height)] = Previewer.Calculate(value, Previewer);
 									}
 								}
